Update checkout customization details
